Fix DnsPolling class
