+++ __tests__/dns-polling.js
@@ -1,0 +1,7 @@
+const { DnsPolling } = require('..');
+
+describe('DnsPolling', () => {
+  it('instantiates', () => {
+    const polling = new DnsPolling();
+  });
+});
